+++ CHANGELOG.md
@@ -23,6 +23,8 @@ All notable changes to this project that require documentation updates will be d
   and the `includeSnoozed` query parameter respectively can be used to include all CVEs.
 - The 'namespace.metadata.stackrox.io/id' label is now removed in order to better support Terraform cluster management.
 - UI: Hovering over a deployment in the Network Graph will show the ports and protocols for it's ingress/egress network flows (ROX-5226).
+- Adding the annotation `auto-upgrade.stackrox.io/preserve-resources=true` on the `sensor` deployment and the `collector` daemonset
+  will cause the auto-upgrader to preserve any overridden resource requests and limits whenever an upgrade is performed.
 
 ## [46.0]
 - Added the following REST APIs: